Removed saving of monitor key
As confirmed by ceph team, the keyring is not the actual
monitor secret to be used while expansion of cluster.

Rather we should save the monitor secret while cluster creation
flow and use the same in expansion time. Will send another PR
for saving the correct monitor secret as part of create cluster flow.

Signed-off-by: Shubhendu <shtripat@redhat.com>