Corrected utilization sync issue
Earlier utilization data for the cluster and pool was being
synchronized on osd_map change only. If we are just populating
objects in pools to increase utilization, the osd_map version is
not going to change and so the utilization will not trigger at all.

Moved the utilization syncing logic outside now and it would be done
triggered after cluster data sync flow.

tendrl-bug-id: Tendrl/ceph-integration#271
Signed-off-by: Shubhendu <shtripat@redhat.com>